+++ src/main/java/duke/task/Deadline.java
@@ -1,4 +1,4 @@
-package task;
+package duke.task;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
